Rename providers to middleware and move code from validator.ts to middleware.ts
